requirements.txt
# ═══════════════════════════════════════════════════════════
# SMA CROSS VN30 - REQUIREMENTS
# ═══════════════════════════════════════════════════════════

# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data Source
vnstock==0.2.9.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dashboard
streamlit>=1.25.0

# Technical Indicators
ta-lib>=0.4.0  # Optional: requires TA-Lib C library
pandas-ta>=0.3.14b0

# Statistics & ML
scipy>=1.10.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL (optional)

# Logging & Monitoring
colorlog>=6.7.0
tqdm>=4.65.0

# CLI
click>=8.1.0
typer>=0.9.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Optimization
numba>=0.57.0  # JIT compilation for speed
joblib>=1.3.0  # Parallel processing

# Utils
python-dotenv>=1.0.0
requests>=2.31.0

